V0.2.1_增加esp32c3深度睡眠

+++ components/MG90S/include/MG90S.h
@@ -2,7 +2,7 @@
  * @Author: shadow MrHload163@163.com
  * @Date: 2023-12-21 09:29:04
  * @LastEditors: shadow MrHload163@163.com
- * @LastEditTime: 2023-12-27 17:31:40
+ * @LastEditTime: 2023-12-28 09:02:23
  * @FilePath: \SmartLock\components\MG90S\include\MG90S.h
  * @Description:
  */
@@ -21,7 +21,7 @@
 typedef struct
 {
     char *tag;
-    uint8_t state;
+    int state;
     TaskHandle_t taskHandle;
     EventGroupHandle_t eventHandle;
 } MG90S_TypeDef, *PMG90S_TypeDef;
